# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel-TODO App",
        "version": "0.1"
    },
    "paths": {
...
        "/api/task/create": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new task",
                "operationId": "73cb494bc7c559a8ccd68efe85b2c6c2",
                "requestBody": {
                    "description": "Task data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "description",
                                    "status",
                                    "due_date",
                                    "category_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "User's id",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "title": {
                                        "description": "task title",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "task description",
                                        "type": "string",
                                        "minLength": 255
                                    },
                                    "status": {
                                        "description": "task status",
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "description": "task due date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "category_id": {
                                        "description": "category id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/update": {
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "Update a task",
                "operationId": "a2e18490e0aff9e234b6a9a9c8de70f9",
                "requestBody": {
                    "description": "Task data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "user_id",
                                    "title",
                                    "description",
                                    "status",
                                    "due_date",
                                    "category_id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Task id",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "user_id": {
                                        "description": "User's id",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "title": {
                                        "description": "task title",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "task description",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "status": {
                                        "description": "task status",
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "description": "task due date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "category_id": {
                                        "description": "category id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/delete": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete a task",
                "operationId": "1feaa77d8b7aa40b12de0ef3bdaa9f94",
                "requestBody": {
                    "description": "Task data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Task id",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/taskList": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get all tasks",
                "operationId": "e9cfcbc75683b4eac5d685fc5cd21663",
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "app/Http/Controllers/TaskController"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred while fetching tasks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}